# src/main/java/com/example/demo/request/Passenger.java
package com.example.demo.request;



public class Passenger {
	private String name;
	private Integer age;
	private String from;
	private String to;
	private String date;
	public String getName() {
		return name;
	}
	public void setName(String name) {
		this.name = name;
	}
	public Integer getAge() {
		return age;
	}
	public void setAge(Integer age) {
		this.age = age;
	}
	public String getFrom() {
		return from;
	}
	public void setFrom(String from) {
		this.from = from;
	}
	public String getTo() {
		return to;
	}
	public void setTo(String to) {
		this.to = to;
	}
	public String getDate() {
		return date;
	}
	public void setDate(String date) {
		this.date = date;
	}
	@Override
	public String toString() {
		return "Passenger [name=" + name + ", age=" + age + ", from=" + from + ", to=" + to + ", date=" + date + "]";
	}
	

}